itemVM refactored with testing class barebone
